+++ BuildMonitor.ADO/ADOBuildStoreFactory.cs
@@ -1,0 +1,12 @@
+﻿using BuildMonitor.Core;
+
+namespace BuildMonitor.ADO
+{
+    public class ADOBuildStoreFactory : IBuildStoreFactory
+    {
+        public IBuildStore GetBuildStore(IMonitorOptions options)
+        {
+            return new ADOBuildStore(options);
+        }
+    }
+}
