--- BuildMonitor.TfsOnline/TfsOnlineBuildStoreFactory.cs
@@ -1,12 +1,0 @@
-﻿using BuildMonitor.Core;
-
-namespace BuildMonitor.TfsOnline
-{
-    public class TfsOnlineBuildStoreFactory : IBuildStoreFactory
-    {
-        public IBuildStore GetBuildStore(IMonitorOptions options)
-        {
-            return new TfsOnlineBuildStore(options);
-        }
-    }
-}
