+++ BuildMonitor.sln
@@ -7,8 +7,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BuildMonitor.Core", "BuildM
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BuildMonitor.UI", "BuildMonitor.UI\BuildMonitor.UI.csproj", "{A210FEC6-E551-4017-9FA3-686C35E040F7}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BuildMonitor.TfsOnline", "BuildMonitor.TfsOnline\BuildMonitor.TfsOnline.csproj", "{880B2245-47CD-4F52-A1AC-C014A87CADFC}"
-EndProject
 Project("{930C7802-8A8C-48F9-8165-68863BCCD9DD}") = "BuildMonitor.Setup", "BuildMonitor.Setup\BuildMonitor.Setup.wixproj", "{8F11FCD3-8C80-4937-87A9-D756B4A77103}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BuildMonitor.TestApp", "BuildMonitor.TestApp\BuildMonitor.TestApp.csproj", "{348A9F74-0A5A-43F3-BEDD-E8AA871A099A}"
@@ -25,6 +23,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 	EndProjectSection
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BuildMonitor.App", "BuildMonitor.App\BuildMonitor.App.csproj", "{8EBC2911-7950-4E5E-90E2-31EF866E8E89}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BuildMonitor.ADO", "BuildMonitor.ADO\BuildMonitor.ADO.csproj", "{8409A0C6-7D7A-B907-5A8C-16708C64D800}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -50,14 +50,6 @@ Global
 		{A210FEC6-E551-4017-9FA3-686C35E040F7}.Release|Any CPU.Build.0 = Release|Any CPU
 		{A210FEC6-E551-4017-9FA3-686C35E040F7}.Release|x86.ActiveCfg = Release|Any CPU
 		{A210FEC6-E551-4017-9FA3-686C35E040F7}.Release|x86.Build.0 = Release|Any CPU
-		{880B2245-47CD-4F52-A1AC-C014A87CADFC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{880B2245-47CD-4F52-A1AC-C014A87CADFC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{880B2245-47CD-4F52-A1AC-C014A87CADFC}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{880B2245-47CD-4F52-A1AC-C014A87CADFC}.Debug|x86.Build.0 = Debug|Any CPU
-		{880B2245-47CD-4F52-A1AC-C014A87CADFC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{880B2245-47CD-4F52-A1AC-C014A87CADFC}.Release|Any CPU.Build.0 = Release|Any CPU
-		{880B2245-47CD-4F52-A1AC-C014A87CADFC}.Release|x86.ActiveCfg = Release|Any CPU
-		{880B2245-47CD-4F52-A1AC-C014A87CADFC}.Release|x86.Build.0 = Release|Any CPU
 		{8F11FCD3-8C80-4937-87A9-D756B4A77103}.Debug|Any CPU.ActiveCfg = Debug|x86
 		{8F11FCD3-8C80-4937-87A9-D756B4A77103}.Debug|Any CPU.Build.0 = Debug|x86
 		{8F11FCD3-8C80-4937-87A9-D756B4A77103}.Debug|x86.ActiveCfg = Debug|x86
@@ -82,6 +74,14 @@ Global
 		{8EBC2911-7950-4E5E-90E2-31EF866E8E89}.Release|Any CPU.Build.0 = Release|Any CPU
 		{8EBC2911-7950-4E5E-90E2-31EF866E8E89}.Release|x86.ActiveCfg = Release|Any CPU
 		{8EBC2911-7950-4E5E-90E2-31EF866E8E89}.Release|x86.Build.0 = Release|Any CPU
+		{8409A0C6-7D7A-B907-5A8C-16708C64D800}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8409A0C6-7D7A-B907-5A8C-16708C64D800}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8409A0C6-7D7A-B907-5A8C-16708C64D800}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{8409A0C6-7D7A-B907-5A8C-16708C64D800}.Debug|x86.Build.0 = Debug|Any CPU
+		{8409A0C6-7D7A-B907-5A8C-16708C64D800}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8409A0C6-7D7A-B907-5A8C-16708C64D800}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8409A0C6-7D7A-B907-5A8C-16708C64D800}.Release|x86.ActiveCfg = Release|Any CPU
+		{8409A0C6-7D7A-B907-5A8C-16708C64D800}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
